Fix script to copy library into usage directory
In Python we can't import from a parent directory, so we can't have a
pymk.py inside usages that imports from the sourcecode at src. I still
want to make pymk.py and mkpy be at the root, so I prefer to use an
internal_scripts.py file to do this and sacrifice autocomplete on it.