Stop calling functions in pymk.py targets, they are now snips
The word target got overloaded when implementing the recipe concept to
the module. To avoid the conflict we now call functions in pymk.py snips
as shorthand of snippet. This commit makes that change and updates the
documentation and function names.

There is an issue because the variable where we stored the last called
function was called last_target. Its now called last_snip, but this
means default calls will break until users update their scripts. I added
a warning describing the change that needs to be made.

Also, pymk_default() now receives an optional list argument that
contains a list of snip names that will not update the last_snip
variable. This is useful for targets that are usually called just once
during development like searches. You want to build, search something
then build again without havinf to pass the snip name as argument each
time.