Use symbolic links for pymk installation in usages
This is much easier to maintain as I will always edit the maín source
file, even when debugging usages. Before I would have to either edit the
copy of it inside the usage or change directories to get to the main
source file, which is inconvenient.

Instead I added a script that automatically generates a usage folder for
a provided name.

I also removed the internal_scripts.py, what I had here is better
located at the pymk.py script inside the usages folder. This also cleans
the root folder, and stops using the library in weird ways (calling
pymk.py differently).