Consistent parsing of remaining arguments from CLI
Before, we implicitly assumed a command that wanted to get the "rest" of
CLI arguments, would have at least one option starting with "-". This
was problematic because it led to inconsistent behavior about inclusion
of the snip's name in the resulting array. The new behavior keeps track
of what's been caught by get_cli_arg() and get_cli_arg(), then the
returned array will contain everything after the first uncaught option
or everything after the first argument not starting with "-".

I also added a new mechanism for marking a snip to not be added to the
cache. This functionality was there before, but it was brittle because
it relied on adding the snip's function name to an array, so you would
need to remember to update the name in 2 places if it changed. The new
approach uses a dedcorator function and avoids having to repeat the name
of the function in multiple places.