Cleanup of command execution functions
Explain the rationale behing not allowing shell syntax in ex_bg().

Add a log parameter to ex_bg() so users can specify a filename where
stdout and stderr will be written to.

Finally remove get_user_str_vars() and all usages of it. This
functionality was limited to substitution of global variables and it's
now obsolete since the introduction of f-strings. Which are much more
flexible.