Add an example of how to create a toggleable server
I've nor refactored how we do usages, having a second pymk.py inside
usages, does not work... I did not even rmemember about it because it's
not visible at first glance. Instead add a meta.py file, which is
actually a pymk.py file for this repository with snips for all
maintainership tasks of the project.