Improve tab completions
Fix dection of execution under interactive environments. Using `echo $-`
only works from bash, because it was being called inside ex(), this
property wasn't being set so it was always in non interactive mode. I
hpe the new approach works better, but I still haven't run the scripts
that were causing the issue of showing the completion warning when it's
unnecessary.

Also fixed the bash code for the completion, now the fallback file name
completion does work as expected. Spaces are escaped and there is no
space after a completion is accepted.